Support import into ES6 modules
With this change, both LineColumn and LineColumnFinder are
accessible from an ES6 module, as follows:

    import licofi from 'licofi'

    let finder = new licofi.LineColumnFinder(...)

    let licol = new licofi.LineColumn(...)

Because LineColumn was declared as the default export, it was
impossible to import LineColumnFinder into an ES6 module.
It didn't make sense that LineColumn was the default export
anyway. See https://nodejs.org/api/esm.html#esm_import_statements